Backend/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
asgiref<4,>=3.5.2
tzdata
pytz
sqlparse>=0.2.2
Pillow>=9.3.0
djangorestframework-simplejwt>=5.2.2
python-dateutil>=2.8.2
